Move algebra package to //common/math

+++ common/lib/math/sign.h
@@ -1,6 +1,5 @@
 #pragma once
 
-template <typename T> int64_t sgn( T val )
-{
+template <typename T> int sgn( T val ) {
     return ( T( 0 ) < val ) - ( val < T( 0 ) );
 }
